Added a UI library so we can have a place to store our custom classes.
stuck in an class that extends DGV to have an image column.  It works except that I don't have any pictures for it yet.

got it here:  http://www.informit.com/articles/article.aspx?p=446453&seqNum=14

as it is just an example, I will later rewrite that code to make it into a "ChessSquare" class.  Basically it will just be a square that displays an image (although later I think I'd like it to have a 'tooltip' property (if it doesn't already have one') so that we can have some rules for tutorials, or messages to appear regarding particular chessSquares.

I assume that there will be a layout() function that will throw down these squares in a calculated fashion (according to the board definition in the rules file) -- size, color, etc..