abandoning the DGV idea.  It won't work for our needs.
Instead I will create a custom grid on a form made up of customized pictureboxes (with Board.Square, and IConfigurablePiece props added)

Both Board.CreateBoard() and this UI form need to get their values from an XML file.

and.. it would be nice if the algorithm that creates the board in Engine.Board.CreateBoard() could also be used to create the UI board. (perhaps by raising an event). - also, rows, colums, etc need to be pulled from XML

I don't want to have to have 2 copies of code rendering each board (as we plan to support multiple boards.)