UI is now pulling its piece data from the Engine
I still haven't worked out who is going to be using AttackingPage and MovementPackage  (right now this is done in Rules, but they are needed in Engine.  It is starting to look like the Rules Library and Engine Library are going to need to be merged to carry it out.

That is fine, we can simply transfer all the rules classes over. It shouldn't be a hard move.