refs #8,
continuing Drag & Drop stuff.

My overridden "AllowDrop" works, so a piece can now be dropped onto another picturebox, and it is deleted in the old one.

I think to show the piece "dragging", that I need to set a "custom cursor", but I'll do that another day.

Done for the night.