refs #30,
so.. apparently I screwed up the handlers for the UI when I threw down the menu bar.  This should fix it..


refs #26,
in other news, the Jabber library was pissing me off because it was getting around my global filter and wanting to check in dlls.

I redid all of the output directories to fix that, which meant that I also had to change around the refs.  Now each jabber project is ref'ing a project rather than pointing to the dll (this way I can change output directories and not break things!)