closes #69,
Everything has been moved out of the UI Form.  The purpose of ChessGrid2D_Form is to provide an eventbase (for buttons, menus, etc) in which to call the ChessMangler object model. 

Created Grid2D to handle all 2D Grid functionality (and most importantly, to make it easier to test! (there are no plans to fake button presses, and menu clicks, just the object model (including grid) will be tested.)