closes #81,
There we go.  Pieces now resize correctly.

The goal here was to provide a little enlargement of the piece showing that it was picked up.  This functionality is intended to be configured in the program and game config files. 

(if the config node is not there in the game config, then it will default back to the setting in the program config)