refs #85, #86,
added some debug code so i can figure out how to fix the grids.

but the debugform is broken.  Did some fixing but it isn't fixed yet.

Added a Tool Form (see #72, #73) so we can start building a grid of spaces in which to stick piece images  

The piece images then will be able to be dragged & dropped onto the Chess Grid