refs #85,
moving some things around so I can manipulate the grid better.

The square handlers should not have been set in the base class (as we need to be able to reset them at some points.. (especially for debug purposes!) 

Added some extra functions (like Delete_Required_Square_Handlers() for example)  that should assist in resetting the grid.

These are extra functions that will be needed for other things later, but they are helping me to visualize things right now so that I can sync the UIGrid w/ the engine grid.