refs #26,
threw down a combobox on the start form to layout some comms stuff.

refs #72,

The bulk of the work was done on the right-click menu to be able to add a piece from any config file that currently resides in the 'config' directory.

in free-form mode, when you right-click on any square, then you can bring up any piece from any config file. (the menu is populated on the fly)  As each piece will eventually contain its own rules, it could make for some interesting games indeed!

for the next part of this (after a refactor), I would like to put the piece images right next to those generated piece menu items as well.

With this commit, we bring xml into the WinUIParts library.

some links:

do it manually:  http://www.codeproject.com/KB/menus/cs_menus.aspx

or using a class that someone wrote:

http://www.codeproject.com/KB/menus/MenuExtender.aspx