closes #72,
Done. As far as the colors, chose a simple implementation for now.  However, laid out plans for a more complicated one (in UniquePieceMenu) if the need arises that we need to do some more advanced searching. (Which is likely if we need to be able to support all sorts of weird stuff in the future)

Also got a freebie as far as the "Transform Piece" menu.  Its the same thing as adding a piece, so we just call the UniquePieceMenu again!

Something worth mentioning:  *All available XML config files are being parsed on right-click*..  This means if a user were to have a whole boatload of config files in the config directory, then they may take a long time to parse.  So I created ticket #93 to deal with this issue.