closes #66,
ok. Have proof of concept for chat. Target is currently hardcoded as we don't have a config set up yet..

However, having cross-thread errors when posting back to the textbox. (RCV uses a MessageBox for the moment..)

That is fine. My design is such that I wanted Jabber to do all of its work elsewhere anyway.

So, to fix this (see #109), There will be a timer on the ChessGrid form (or GridForm) that will instantiate a "Jabber" object and poll a "Recvd Message" prop on it to display the latest recieved message

That object will also manage the "Message Buffer" (which personally, I think should be saved to the SQLite database as well ~ in a dedicated "Chat" Table.

also, the fact that the ChatBox does not resize with the Form is getting annoying.  Fix this.