refs #121,
Laid out MenuItems (this will be refactored like the previous menuItem sets)

For now we are just going to build (on the fly) a list of available 'piece image' sets for the user to pick from (and if any pieces don't line up, then they just won't be replaced)

Later (> .5) or (>1.0) it would be supremely nice (when you hover over an image set) to have it pop out a "preview" of the pieces, or even slap images of all the pieces in the menuitem (although I wonder how big that could make the menuItem ~ a preview with fuzzy borders off to the side would probably do the job too)

also, removed the opponent hardcode in Grid2D form and moved it to the gameList.  (Laid out some add'l pseudocode for that 'opponents' cboBox) (see #122)

also, the main form has the name of the Application on it now.