refs #42,
Added the Move "OutBox" concept.  Now when you make a move, the "Move Packet" (represented as an XMLElement) will be added to the OutBox. "Submit" will submit the latest move in the OutBox.

refs #97,

starting down this direction too.  Threw down some initial stuff. It needs to be cleaned up.

Also:  UIBoard.Squares now derives from a new object that inherits List<square> as we are going to have a number of operations that will need to be done on all UISquares in the grid.