refs #109,
hmmmmm.  Not sure if I want to go through all of the effort on this one.  The Jabber Control wants to be on the form, and making it all work in its own "Handlers" class seems like a lot of work just to strip the code out of the form. (Implementing ISynchronizeInvoke). I really don't want to have to write all of that code if I don't have to..

(hmmmmm. come to think of it.. one easy way out of it would be to have JabberHandlers inherit System.Windows.Forms.Control..  That would save having to reimplement..)