refs #103,
Laid the groundwork for this.  

The ChessGrid is set up so that it is no longer "hardcoded" to a particular "Chat Provider". yeah, I could have worded this ticket better, but I know what I mean  :D 

Now all you need to do is to swap out the "Comms Type" when adding the "CommsHandlers" to the form.