refs #140,
All of this was just to get access to the ICommsHandler.On_Authenticate handler.

It just so happens that with the interface set up the way it is, that this is the only way it can be done at present (Later I will need to revise the Comms interface.)

We need access to the On_Authenticate handler on the GameList so that the user can login and get a list of opponents to play.