refs #140,
Some cleanup of the UI form.  Now that we have the "Authenticated" event, we have a point in which to populate the roster so the user can select an opponent.

As before. The way that you currently select an opponent is to type it in manually in the text box. (however, that will soon go away)