refs #140,
it was the cross-thread issues.  unfortunately, the muzzle.rosterUIThingy aint working no matter what I do.. Checking this in before I rip it all out.

I *was* going to try and tailor this roster UI thingy, but I really wanted to have a datagrid anyway. not only that, it is Jabber-only, so it couldn't stay anyway..

now I can get the JID stuff, so the next step is to:

1. Make the users visible in a grid that will do what the rosterThingy did

2. Refactor all this crapola code into the interface.. its all Jabber-specific.

3.  The new grid will probably wind up as its own custom control anyway..