refs #149,
its working on the chessgrid form, but only because it is being passed in from the gamelist.  I tried doing it just like I did in the gamelist, but "this.Components" is always null!

Now that I think about it, that might be likely to my inherited forms.  I might jsut have needed to ref the base form or something..

Whatever the case, it needs to be set in the ICommsHandler, and this ticket *should* be closed when it is refactored to there.

We will see what I think later.. if tthis ticket is closed without the refactoring, then make it have a refactor ticket