refs #144, #161
Moved ExecuteMove to ChessMangler.Communications.

What does this have to do with Implementing version? Well this has more about changing the architecture to accomodate it, as the engine does not need to be worrying about comms. It need only worry about moving stuff around.

Now I'll be able to use the Engine (and its Xml parsing abilities) from the Comms layer, (which was not possible before due to a circular ref. conflict))