refs #180,
Project is now using Visual Studio 2010 Beta 2.

It is still using .net 3.5 as, although NUnit claims that it will work with .net 4.0, the testRunner doesn't seem to work with it yet..

Also, project now uses the latest version of NUnit  (2.5.3.9325)

Gallio still works fine but will need to be upgraded once we upgrade to .net 4.0.

Gallio appears to work fine in Win 7 as long as its ren as administrator.